requirements.txt
pandas
llama_index.llms.openai
llama_index.core
openpyxl
streamlit
requests